+++ jdaviz/app.vue
@@ -87,6 +87,7 @@
                   @resize="relayout"
                   :closefn="destroy_viewer_item"
                   @data-item-selected="data_item_selected($event)"
+                  @data-item-visibility="data_item_visibility($event)"
                   @data-item-remove="data_item_remove($event)"
                   @call-viewer-method="call_viewer_method($event)"
                 ></g-viewer-tab>