+++ jdaviz/components/tooltip.vue
@@ -61,8 +61,10 @@ const tooltips = {
   'viewer-toolbar-more': 'More options...',
   'viewer-data-select-enabled': 'Allow multiple entries (click to enable replace)',
   'viewer-data-radio-enabled': 'Replace current entry (click to enable multi-select)',
-  'viewer-data-select': 'Toggle whether data entry is loaded in the viewer',
-  'viewer-data-radio': 'Change viewer to this data entry',
+  'viewer-data-select': 'Toggle visibility of all layers associated with this data entry',
+  'viewer-data-radio': 'Switch visibility to layers associated with this data entry',
+  'viewer-data-enable': 'Load data entry into this viewer',
+  'viewer-data-disable': 'Disable data within this viewer (will be hidden and unavailable from plugins until re-enabled)',
   'viewer-data-delete': 'Remove data entry across entire app',
 
   'table-prev': 'Select previous row in table',