+++ jdaviz/components/viewer_data_select_item.vue
@@ -1,20 +1,25 @@
 <template>
   <div style="display: contents">
-    <div>
-    <j-tooltip :tipid="multi_select ? 'viewer-data-select' : 'viewer-data-radio'">
-      <v-btn 
-        icon
-        :color="isSelected ? 'accent' : 'default'"
-        @click="$emit('data-item-selected', {
-          id: viewer.id,
-          item_id: item.id,
-          checked: !isSelected,
-          replace: !multi_select
-        })">
-          <v-icon v-if="multi_select">{{isSelected ? "mdi-checkbox-marked" : "mdi-checkbox-blank-outline"}}</v-icon>
-          <v-icon v-else>{{isSelected ? "mdi-radiobox-marked" : "mdi-radiobox-blank"}}</v-icon>
-      </v-btn>
-    </j-tooltip>
+    <div v-if="isSelected">
+      <j-tooltip :tipid="multi_select ? 'viewer-data-select' : 'viewer-data-radio'">
+        <v-btn 
+          icon
+          :color="visibleState==='visible' ? 'accent' : 'default'"
+          @click="selectClicked">
+            <v-icon v-if="multi_select">{{visibleState!=='hidden' ? "mdi-checkbox-marked" : "mdi-checkbox-blank-outline"}}</v-icon>
+            <v-icon v-else>{{visibleState!=='hidden' ? "mdi-radiobox-marked" : "mdi-radiobox-blank"}}</v-icon>
+        </v-btn>
+      </j-tooltip>
+    </div>
+    <div v-else>
+      <j-tooltip tipid="viewer-data-enable">
+        <v-btn 
+          icon
+          color="default"
+          @click="selectClicked">
+            <v-icon>mdi-plus</v-icon>
+        </v-btn>
+      </j-tooltip>
     </div>
 
     <j-tooltip :tooltipcontent="'data label: '+item.name" span_style="font-size: 12pt; padding-top: 6px; padding-left: 6px; width: calc(100% - 80px); cursor: default;">
@@ -28,8 +33,21 @@
       </div>
     </j-tooltip>
 
+    <div v-if="isSelected && isUnloadable" style="position: absolute; right: 5px">
+      <j-tooltip tipid='viewer-data-disable'>
+        <v-btn
+          icon
+          @click="$emit('data-item-selected', {
+            id: viewer.id,
+            item_id: item.id,
+            checked: false,
+            replace: false
+          })"
+        ><v-icon>mdi-close</v-icon></v-btn>
+      </j-tooltip>
+    </div>
 
-    <div v-if="isDeletable" style="position: absolute; right: 10px">
+    <div v-if="isDeletable" style="position: absolute; right: 5px">
       <j-tooltip tipid='viewer-data-delete'>
         <v-btn
           icon
@@ -44,6 +62,32 @@
 
 module.exports = {
   props: ['item', 'multi_select', 'viewer'],
+  methods: {
+    selectClicked() {
+      prevVisibleState = this.visibleState
+
+      if (!this.isSelected) {
+        // data currently isn't loaded, so clicking will LOAD the data (and become visible immediately)
+        this.$emit('data-item-selected', {
+          id: this.$props.viewer.id,
+          item_id: this.$props.item.id,
+          checked: true,
+          replace: false
+        })
+      }
+
+      // checkboxes control VISIBILITY of layers not loaded state
+      // if we just loaded the data, its probably already visible, but we'll still make sure all
+      // appropriate layers are visible and properly handle replace for non-multiselect
+      this.$emit('data-item-visibility', {
+        id: this.$props.viewer.id,
+        item_id: this.$props.item.id,
+        visible: prevVisibleState != 'visible' || !this.multi_select,
+        replace: !this.multi_select
+      })
+
+    }
+  },
   computed: {
     itemNamePrefix() {
       if (this.$props.item.name.indexOf("[") !== -1) {
@@ -62,8 +106,31 @@ module.exports = {
       }
     },
     isSelected() {
-      return this.$props.viewer.selected_data_items.includes(this.$props.item.id)
-    }, 
+      return Object.keys(this.$props.viewer.selected_data_items).includes(this.$props.item.id)
+    },
+    visibleState() {
+      return this.$props.viewer.selected_data_items[this.$props.item.id] || 'hidden'
+    },
+    isUnloadable() {
+      if (this.$props.item.meta.Plugin !== undefined) {
+        return true
+      }
+      if (this.$props.viewer.config === 'cubeviz') {
+        // forbid unloading the original reference cube
+        // this logic might need to be generalized if supporting custom data labels
+        // per-cube or renaming data labels
+        if (this.$props.viewer.reference === 'flux-viewer') {
+          return this.$props.item.name.indexOf('[FLUX]') === -1
+        } else if (this.$props.viewer.reference === 'uncert-viewer') {
+          return this.$props.item.name.indexOf('[IVAR]') === -1
+        } else if (this.$props.viewer.reference === 'mask-viewer') {
+          return this.$props.item.name.indexOf('[MASK]') === -1
+        } else if (this.$props.viewer.reference === 'spectrum-viewer') {
+          return this.$props.item.name.indexOf('[FLUX]') === -1          
+        }
+      }
+      return true
+    },
     isDeletable() {
       // only allow deleting products from plugins.  We might want to allow some non-plugin
       // data to also be deleted in the future, but would probably need more advanced logic
@@ -72,8 +139,22 @@ module.exports = {
         return false
       }
       // for any exceptions not above, enable deleting
-      return true
+      return !this.isSelected
     },
+    selectTipId() {
+      if (this.multi_select) {
+        return 'viewer-data-select'
+      } else {
+        return 'viewer-data-radio'
+      }
+    },
+    labelColor() {
+      if (this.isSelected) {
+        return 'black'
+      } else {
+        return 'gray'
+      }
+    }
   }
 };
 </script>