+++ jdaviz/configs/default/plugins/plot_options/plot_options.py
@@ -27,8 +27,8 @@ class PlotOptions(TemplateMixin):
     layer_selected = Any().tag(sync=True)  # Any needed for multiselect
 
     # spectrum viewer/layer options:
-    line_visible_value = Bool().tag(sync=True)
-    line_visible_sync = Dict().tag(sync=True)
+    layer_visible_value = Bool().tag(sync=True)
+    layer_visible_sync = Dict().tag(sync=True)
 
     collapse_func_value = Unicode().tag(sync=True)
     collapse_func_sync = Dict().tag(sync=True)
@@ -121,10 +121,13 @@ class PlotOptions(TemplateMixin):
         def is_spatial_subset(state):
             return isinstance(state, ImageSubsetLayerState)
 
+        def is_not_subset(state):
+            return not isinstance(state, ImageSubsetLayerState)
+
         # Spectrum viewer/layer options:
-        self.line_visible = PlotOptionsSyncState(self, self.viewer, self.layer, 'visible',
-                                                 'line_visible_value', 'line_visible_sync',
-                                                 state_filter=is_profile)
+        self.layer_visible = PlotOptionsSyncState(self, self.viewer, self.layer, 'visible',
+                                                  'layer_visible_value', 'layer_visible_sync',
+                                                  state_filter=is_not_subset)
         self.collapse_function = PlotOptionsSyncState(self, self.viewer, self.layer, 'function',
                                                       'collapse_func_value', 'collapse_func_sync')
         self.line_color = PlotOptionsSyncState(self, self.viewer, self.layer, 'color',